Move the `define_env` proc-macro to its own module
